Move rollout controller to separate package
